Major overhaul of the queued chunk-loading
It is finally working correctly and reliable. Actually I am baffled,
that it had worked at all before. Fixing a rounding error in the statement
     chunkCoordinate = (int) playerLocation >> 4
triggered all this. The correct line is:
     chunkCoordinate = MathHelper.floor(playerLocation) >> 4
In the negative range, the faulty line calculates the wrong chunk
coordinates within the first block in the chunk.
In the range ]-1,0[ for example