Found a much simpler solution for concurrent chunk generation
One BiomeDecorator can only decorate one chunk at a time, but we can create shallow copies of that decorator on demand in order to handle multiple chunks at the same time. The IntCache in its static synchronized form is a major burden for performance here. I converted it to multiple thread local ones without the need to synchronize