ACHIEVEMENT GET! Offloaded path searches This change offloads all the path searches for the navigation tasks of the subclasses of EntityInsentient. This lowers the burden from hostile entities for single worlds drastically. Commit 6a75666 was only for a few -oldAI- mobs.
Theres a minor flaw though. The PathFinderGoals dont get back the result for the path search immediatelly. This delays the processing of the goal by a tick or two, depending what time the path search takes. The PathEntity of the Navigation object of an EntityInsentient is updated then once the search is done. I didnt do much testing in 1.7.9 with this yet, but in my 1.7.2 tests all across the board, along the development of this, this issue was barely noticable though, if at all. I also didnt find anything yet that broke with this change in 1.7.2, but that case is not ruled out, neither for 1.7.2, nor for 1.7.9