Fixed disappearing of leash hitches in world spawn areas When an leashed entity is loaded, the information about it being leashed and the leash holder is loaded correctly, though not being processed yet. The leash hitch is set up when the entity is ticked the first time. If a chunk save happens before that, the information about the leash holder is not stored, as none has been created at that point yet.
This fix adds a call to the method bP(), which sets up the leash hitch, if it was not done yet before saving the entity's data
Furthermore the if-elseif-else construct in bP() is split up into separate ifs, as the UUIDMost and UUIDLeast information is always present in the entity's data. It's use is ambiguous. In most places it is used to store the entity's own UUID, and here it is used to store the leash holder's.