Fixed disappearing of leash hitches in world spawn areas When a leashed entity is loaded, the information about it being leashed and the leash holder is loaded correctly, though not being processed yet. The leash hitch is set up when the enity is ticked the first time. If a chunk save happens before that, the information about the leash holder is not stored, as none has been created at that point yet.
This fix adds a call to the method bJ(), which sets up the leash hitch, if it was not done yet before saving the entity's data. Furthermore the if-elseif-else construct in bJ() is split up into separate ifs, as the UUIDMost and UUIDLeast information is always present in the entity's data. It's use is ambigous. In most places it is used to store the entity's own UUID, and here it is used to store the leash holder's.

As saving a leash entity may add a new entity (a leash hitch) to the current world now, the usage of an iterator to go through the list of entities in a chunk slice, is not safe here anymore. The iterator should throw a ConcurrentModificationException in that case, as the list was modified during iteration. Interestingly it didnt do that in my tests, but it should