ACHIEVEMENT GET! Offloaded path searches This change offloads all the path searches for the navigation tasks of the subclasses of EntityInsentient. This lowers the burden from hostile entities for single worlds drastically.
There's a minor flaw though. The PathfinderGoals don't get back the result of the path search immediatelly. This delays the processing of the goals by a tick or two, depending on how much time the path search takes. The PathEntity of the Navigation object of an EntityInsentient is updated thenonce the search is done