Configurable dynamic number of chunks per packet On our new root server, we noticed that the loading and sending of chunks to the players is so much faster now, that the chunk packets hog the entire bandwidth of the client for a few seconds. Especially when the players teleport and the server sends them the 441 chunks (VD 10) at once. Even though they are already split up in packets of 5 chunks each and one packet per tick, it completly blocks all other game packets during that period and causes a very noticable connection lag spike.
This change lets the PlayerChunkManager decide how many chunks each packet should have, depending on the priority of the chunks that are queued next in the PlayerChunkSendQueue. The priority of a chunk is determined by the MinetickChunkCoordComparator depending on the position of the chunk within the area of the player and the player's current viewing direction (yaw). The MinetickChunkCoordComparator's general idea is to priorize chunks that are close and infront of the player over those that are further away and/or behind the player.
There are four priorities defined and for each priority the number of chunks per packet can be defined in the config. The goal there is to send the very close chunks still fast and then drastically slow down sending as the priority of the queued chunks falls